refactor: GameResult 위치를 game 패키지로 이동
